+++ lib/iris/tests/__init__.py
@@ -330,10 +330,6 @@ class IrisTest(unittest.TestCase):
         np.testing.assert_array_equal(a.mask, b.mask)
         np.testing.assert_array_almost_equal(a[~a.mask].data, b[~b.mask].data)
 
-    def assertMaskedArrayEqual(self, a, b):
-        np.testing.assert_array_equal(a.mask, b.mask)
-        np.testing.assert_array_equal(a, b)
-
     def assertArrayAllClose(self, a, b, rtol=1.0e-7, atol=0.0, **kwargs):
         """
         Check arrays are equal, within given relative + absolute tolerances.
